+++ core/core.native/source/core.native/io/binary_reader/STL/serialize_deque.h
@@ -16,22 +16,22 @@ namespace useless
 	struct binary_read_helper<std::deque<Type, Allocator>>
 	{
 		template<typename Archive>
-		static void read_member( Archive& br, std::deque<Type, Allocator>& val, unsigned long count, std::true_type )
+		static void read_member( Archive& br, std::deque<Type, Allocator>& val, uint32_t count, std::true_type )
 		{
 			br.read( &val[ 0 ], sizeof( Type ) * count );
 		}
 
 		template<typename Archive>
-		static void read_member( Archive& br, std::deque<Type, Allocator>& val, unsigned long count, std::false_type )
+		static void read_member( Archive& br, std::deque<Type, Allocator>& val, uint32_t count, std::false_type )
 		{
-			for( unsigned long i = 0; i < count; ++i )
+			for( uint32_t i = 0; i < count; ++i )
 			{
 				br.read( val[ i ] );
 			}
 		}
 
 		template<typename Archive>
-		static void read_member( Archive& br, std::deque<Type, Allocator>& val, unsigned long count )
+		static void read_member( Archive& br, std::deque<Type, Allocator>& val, uint32_t count )
 		{
 			if( !val.empty() )
 			{
@@ -43,10 +43,10 @@ namespace useless
 		template<typename Archive>
 		static void invoke( Archive& br, std::deque<Type, Allocator>& val )
 		{
-			unsigned long current_count = static_cast< unsigned long >( val.size() );
+			uint32_t current_count = static_cast< uint32_t >( val.size() );
 			
-			unsigned long count;
-			br.read( &count, sizeof( unsigned long ) );
+			uint32_t count;
+			br.read( &count, sizeof( uint32_t ) );
 			
 			if( count > current_count )
 			{
