+++ core/core.native/source/core.native/io/binary_reader/STL/serialize_set.h
@@ -16,9 +16,9 @@ namespace useless
 	struct binary_read_helper<std::set<Key, Compare, Allocator>>
 	{
 		template<typename Archive>
-		static void read_member( Archive& br, std::set<Key, Compare, Allocator>& val, unsigned long count )
+		static void read_member( Archive& br, std::set<Key, Compare, Allocator>& val, uint32_t count )
 		{
-			for( unsigned long i = 0; i < count; ++i )
+			for( uint32_t i = 0; i < count; ++i )
 			{
 				Key key;
 				br.read( key );
@@ -29,8 +29,8 @@ namespace useless
 		template<typename Archive>
 		static void invoke( Archive& br, std::set<Key, Compare, Allocator>& val )
 		{
-			unsigned long count;
-			br.read( &count, sizeof( unsigned long ) );
+			uint32_t count;
+			br.read( &count, sizeof( uint32_t ) );
 			read_member( br, val, count );
 		}
     };
@@ -39,9 +39,9 @@ namespace useless
     struct binary_read_helper<std::multiset<Key, Compare, Allocator>>
     {
 		template<typename Archive>
-		static void read_member( Archive& br, std::multiset<Key, Compare, Allocator>& val, unsigned long count )
+		static void read_member( Archive& br, std::multiset<Key, Compare, Allocator>& val, uint32_t count )
 		{
-			for( unsigned long i = 0; i < count; ++i )
+			for( uint32_t i = 0; i < count; ++i )
 			{
 				Key key;
 				br.read( key );
@@ -52,8 +52,8 @@ namespace useless
 		template<typename Archive>
 		static void invoke( Archive& br, std::multiset<Key, Compare, Allocator>& val )
 		{
-			unsigned long count;
-			br.read( &count, sizeof( unsigned long ) );
+			uint32_t count;
+			br.read( &count, sizeof( uint32_t ) );
 			read_member( br, val, count );
 		}
     };
@@ -62,9 +62,9 @@ namespace useless
     struct binary_read_helper<std::unordered_set<Key, Hasher, Compare, Allocator>>
     {
 		template<typename Archive>
-		static void read_member( Archive& br, std::unordered_set<Key, Hasher, Compare, Allocator>& val, unsigned long count )
+		static void read_member( Archive& br, std::unordered_set<Key, Hasher, Compare, Allocator>& val, uint32_t count )
 		{
-			for( unsigned long i = 0; i < count; ++i )
+			for( uint32_t i = 0; i < count; ++i )
 			{
 				Key key;
 				br.read( key );
@@ -75,8 +75,8 @@ namespace useless
 		template<typename Archive>
 		static void invoke( Archive& br, std::unordered_set<Key, Hasher, Compare, Allocator>& val )
 		{
-			unsigned long count;
-			br.read( &count, sizeof( unsigned long ) );
+			uint32_t count;
+			br.read( &count, sizeof( uint32_t ) );
 			read_member( br, val, count );
 		}
     };
