+++ core/core.native/source/core.native/io/binary_writer/STL/serialize_deque.h
@@ -16,22 +16,22 @@ namespace useless
 	struct binary_write_helper<std::deque<Type, Allocator>>
 	{
 		template<typename Archive>
-		static void write_member( Archive& bw, const std::deque<Type, Allocator>& val, unsigned long count, std::true_type )
+		static void write_member( Archive& bw, const std::deque<Type, Allocator>& val, uint32_t count, std::true_type )
 		{
 			bw.write( &val[ 0 ], sizeof( Type ) * count );
 		}
 
 		template<typename Archive>
-		static void write_member( Archive& bw, const std::deque<Type, Allocator>& val, unsigned long count, std::false_type )
+		static void write_member( Archive& bw, const std::deque<Type, Allocator>& val, uint32_t count, std::false_type )
 		{
-			for( unsigned long i = 0; i < count; ++i )
+			for( uint32_t i = 0; i < count; ++i )
 			{
 				bw.write( val[ i ] );
 			}
 		}
 
 		template<typename Archive>
-		static void write_member( Archive& bw, const std::deque<Type, Allocator>& val, unsigned long count )
+		static void write_member( Archive& bw, const std::deque<Type, Allocator>& val, uint32_t count )
 		{
 			if( !val.empty() )
 			{
@@ -43,8 +43,8 @@ namespace useless
 		template<typename Archive>
 		static void invoke( Archive& bw, const std::deque<Type, Allocator>& val )
 		{
-			unsigned long count = static_cast< unsigned long >( val.size() );
-			bw.write( &count, sizeof( unsigned long ) );
+			uint32_t count = static_cast< uint32_t >( val.size() );
+			bw.write( &count, sizeof( uint32_t ) );
 			write_member( bw, val, count );
 		}
     };
